LayoutTests/imported/w3c/web-platform-tests/html/semantics/invokers/interestelement-interface.tentative-expected.txt



FAIL interestTargetElement reflects interestee HTML element assert_equals: expected (object) Element node <div id="interestee"></div> but got (undefined) undefined
FAIL interestTargetElement reflects set value assert_equals: expected "" but got "interestee"
FAIL interestTargetElement reflects set value across shadow root into light dom assert_equals: expected "" but got "interestee"
FAIL interestTargetElement does not reflect set value inside shadowroot assert_equals: expected null but got Element node <div></div>
FAIL interestTargetElement does not reflect invalid value assert_equals: expected null but got Element node <div></div>
FAIL interestTargetElement throws error on assignment of non Element assert_throws_js: interestTargetElement attribute value must be an instance of Element function "function () {
        buttonInvoker.interestTargetElement = {};
      }" did not throw
FAIL interestAction reflects '' when attribute not present assert_equals: expected (string) "" but got (undefined) undefined
FAIL interestAction reflects '' when attribute empty, setAttribute version assert_equals: expected (string) "" but got (undefined) undefined
FAIL interestAction reflects '' when attribute empty, IDL setter version assert_equals: expected (string) "" but got (object) null
FAIL interestAction reflects same casing assert_equals: expected "fooBarBaz" but got ""
FAIL interestAction reflects '' when attribute set to [] assert_equals: expected (string) "" but got (object) []
FAIL interestAction reflects tostring value assert_equals: expected "1,2,3" but got ""
FAIL interestAction reflects tostring value 2 assert_equals: expected (string) "[object Object]" but got (object) object "[object Object]"
